pyproject.toml
[tool.poetry]
name = "python-project-template"
version = "0.1.0"
description = "A template Python project with best practices"
authors = ["Your Name <your.email@example.com>"]
packages = [{ include = "src" }]
readme = "README.md"
license = "MIT"
repository = "https://github.com/yourusername/python-project-template"

[tool.poetry.dependencies]
python = "^3.12"
# Core numeric libraries
jax = { version = "^0.4.26" }
jaxtyping = "^0.3.1"
flax = "^0.8.2"
optax = "^0.2.3"
# Data processing
pandas = "^2.2"
# CLI and configuration
typer = "^0.9"
pyyaml = "^6.0"
python-dotenv = "^1.1.0"
# Visualization
matplotlib = "^3.8"
# Development tools
ipython = "^8.18"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
hypothesis = "^6.97"
ruff = "^0.4"
pre-commit = "^3.7"
nbstripout = "^0.6"
pytest-cov = "^4.1"
pytest-xdist = "^3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "--cov=src --cov-report=term-missing:skip-covered"
markers = [
  "perf: marks tests as performance benchmarks (deselect with '-m \"not perf\"')",
]

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "def __repr__",
  "raise NotImplementedError",
  "if __name__ == .__main__.:",
  "pass",
  "raise ImportError",
]
